Move jest test files out of the expo-router routes directory
index.test.ts and course-timetable/markdownHelpers.test.ts lived inside
app/, so expo-router picked them up as routes. During dev-server SSR the
top-level describe() calls ran outside jest, where describe is undefined,
and every page failed to render with HTTP 500 ("Metro error: describe is
not defined"). Move both files to __tests__/ (jest-expo picks them up
there unchanged) and adjust their relative imports; the route modules
they test stay where they are.
