+++ apps/frontend/app/__tests__/index.test.ts
@@ -1,4 +1,4 @@
-import { extractRawExpoToken } from './index';
+import { extractRawExpoToken } from '../app/index';
 
 describe('extractRawExpoToken', () => {
 	it('extracts the raw token from ExponentPushToken[...] notation', () => {
